# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Lighter Trading Bot API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bot/start/{strategyId}:
    post:
      operationId: startBot
      tags: [bot]
      summary: Start a strategy bot
      parameters:
        - name: strategyId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Bot started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotStatusResponse"
  /bot/stop/{strategyId}:
    post:
      operationId: stopBot
      tags: [bot]
      summary: Stop a strategy bot
      parameters:
        - name: strategyId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Bot stopped
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotStatusResponse"
...
  /bot/pnl-chart:
    get:
      operationId: getPnlChart
      tags: [bot]
      summary: Get PnL chart data for a strategy
      parameters:
        - name: strategyId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: PnL chart data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PnlChartResponse"
  /bot/logs:
    get:
      operationId: getBotLogs
      tags: [bot]
      summary: Get recent bot activity logs
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
      responses:
        "200":
          description: Bot logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotLogsResponse"
components:
  schemas:
...
    BotStatusResponse:
      type: object
      properties:
        strategyId:
          type: integer
        isRunning:
          type: boolean
        message:
          type: string
        nextRunAt:
          type: string
          nullable: true
      required: [strategyId, isRunning, message]
...
    BotLogsResponse:
      type: object
      properties:
        logs:
          type: array
          items:
            $ref: "#/components/schemas/BotLog"
      required: [logs]
    BotLog:
      type: object
      properties:
        id:
          type: integer
        strategyId:
          type: integer
          nullable: true
        strategyName:
          type: string
          nullable: true
        level:
          type: string
          enum: [info, warn, error, success]
        message:
          type: string
        details:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, level, message, createdAt]
    PnlChartResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/PnlChartPoint"
      required: [data]
    PnlChartPoint:
      type: object
      properties:
        date:
          type: string
        buys:
          type: integer
        sells:
          type: integer
        buyVolume:
          type: number
        sellVolume:
          type: number
        estimatedPnl:
          type: number
        cumulativePnl:
          type: number
      required: [date, buys, sells, buyVolume, sellVolume, estimatedPnl, cumulativePnl]